Initial working version.
This is a first version of a brute force word generator. There's no JSLINT/JSHINT validation in there, just a plain generator.

I will "completely" different now as this version can calculate up to only 10-char words on my machine as they integers for counting the possibilities outreaches the limit of integers very quickly (here at around 9 or 10 characters). Therefore i try to achieve a solution that does not need to calculate such huge integers. I will do that somehow with string or arrays, but anyways first i have to save this version.